reorganize weight_ticket_set_documents tests
